! sis900: 32 Bit für eine MAC-Adresse sind einfach zu wenig
git-svn-id: svn+ssh://overgames.de/lost/trunk@799 1fb02b30-9e10-0410-89f7-8f5a202ca6a9

+++ sis900/device.c
@@ -154,7 +154,7 @@ static uint64_t get_mac_address(struct sis900_device* device)
     
     if (device->revision < 0x82) {
     for (i = 0; i < 3; i++) {
-        mac |= sis900_eeprom_read(device, EEPROM_OFS_MAC + i) << (i * 16); 
+        mac |= (uint64_t) sis900_eeprom_read(device, EEPROM_OFS_MAC + i) << (i * 16); 
     }
     } else if ((device->revision >= 0x82) && (device->revision <= 0x90)) {
     // CR_RELOAD_MAC laedt die MAC-Adresse in den Filter
